+++ icicle/src/rings/random_sampling.cpp
@@ -6,10 +6,10 @@ namespace icicle {
 
   ICICLE_DISPATCHER_INST(RandomSamplingRingDispatcher, ring_zq_random_sampling, ringZqRandomSamplingImpl);
   extern "C" eIcicleError CONCAT_EXPAND(ICICLE_FFI_PREFIX, random_sampling)(
-    uint64_t size,
+    size_t size,
     bool fast_mode,
     const std::byte* seed,
-    uint64_t seed_len,
+    size_t seed_len,
     const VecOpsConfig* config,
     field_t* output)
   {
@@ -18,10 +18,10 @@ namespace icicle {
 
   template <>
   eIcicleError random_sampling(
-    uint64_t size,
+    size_t size,
     bool fast_mode,
     const std::byte* seed,
-    uint64_t seed_len,
+    size_t seed_len,
     const VecOpsConfig& config,
     field_t* output)
   {
@@ -34,7 +34,7 @@ namespace icicle {
     challengeSpacePolynomialsSamplingImpl);
   extern "C" eIcicleError CONCAT_EXPAND(ICICLE_FFI_PREFIX, sample_challenge_space_polynomials)(
     const std::byte* seed,
-    uint64_t seed_len,
+    size_t seed_len,
     size_t size,
     uint32_t ones,
     uint32_t twos,
@@ -48,7 +48,7 @@ namespace icicle {
   template <>
   eIcicleError sample_challenge_space_polynomials(
     const std::byte* seed,
-    uint64_t seed_len,
+    size_t seed_len,
     size_t size,
     uint32_t ones,
     uint32_t twos,
